extracted ops data
Ops data (log and pid files) was stored in the app directory,
which is compleetely destroyed and rebuilt during every deploy.
In order to get the ops data to be persistent, it has been moved.

Previous layout:
/home/deploy/apps/web_admin           -> application
/home/deploy/apps/web_admin/logs      -> log files
/home/deploy/apps/web_admin/tmp/pids  -> pid files

New layout:
/home/deploy/apps/web_admin  -> application
/home/deploy/logs/web_admin  -> log files
/home/deploy/pids/web_admin  -> pid files

During the deploy process, we now create symlinks to the log
and pid directories.